requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
httpx>=0.27,<1
python-dotenv>=1.0,<2
pydantic>=2.5,<3
